backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
# Security
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0
python-jose[cryptography]>=3.3.0
# Rate limiting
slowapi>=0.1.9
# Email
aiosmtplib>=3.0.0
# AI/Embeddings
openai>=1.12.0
pgvector>=0.2.4
requests>=2.31.0
# Templates
jinja2>=3.1.0
# Telegram Bot
python-telegram-bot[job-queue]>=21.0
